// src/app/main/pages/compartidos/interfaces/Iinstitucion.ts
export interface Iinstitucion {
    codInstitucionEducativa: number;
    codAmie: string;
    nomInstitucionEducativa: string;
    nomEstadoIe: string;
    nomSostenimiento: string;
    descripcionRegimen: string;
    descripcionJurisdiccion: string;
    nomZona: string;
    nomDistrito: string;
    codAdDistrito: string;
    nomCircuito: string;
    codAdCircuito: string;
    nomProvincia: string;
    nomCanton: string;
    codDpaParroquia: string;
    nomParroquia: string;
}